basic focus overhaul

+++ sway/commands/kill.c
@@ -10,11 +10,16 @@ struct cmd_results *cmd_kill(int argc, char **argv) {
 		return cmd_results_new(CMD_FAILURE, "kill",
 			"Command 'kill' cannot be used in the config file");
 	}
+	if (config->handler_context.current_container == NULL) {
+		wlr_log(L_DEBUG, "no container to kill");
+		return cmd_results_new(CMD_SUCCESS, NULL, NULL);
+	}
 	enum swayc_types type = config->handler_context.current_container->type;
-	if (type != C_VIEW || type != C_CONTAINER) {
+	if (type != C_VIEW && type != C_CONTAINER) {
 		return cmd_results_new(CMD_INVALID, NULL,
 				"Can only kill views and containers with this command");
 	}
+
 	// TODO close arbitrary containers without a view
 	struct sway_view *view =
 		config->handler_context.current_container->sway_view;
